xml resouce

+++ core/src/main/java/org/itas/core/EnumInt.java
@@ -3,6 +3,6 @@ package org.itas.core;
 
 public interface EnumInt {
 
-	public byte key();
+	public int key();
 	
 }
